--- a/bpf/Makefile
+++ b/bpf/Makefile
@@ -96,8 +96,8 @@ load-map-local_pod_ips:
 load-map-process_ip:
 	[ -f $(PROG_MOUNT_PATH)/process_ip ] || sudo bpftool map create $(PROG_MOUNT_PATH)/process_ip type lru_hash key 4 value 4 entries 1024 name process_ip
 
-load-map-cgroup_ips:
-	[ -f $(PROG_MOUNT_PATH)/cgroup_ips ] || sudo bpftool map create $(PROG_MOUNT_PATH)/cgroup_ips type lru_hash key 8 value 16 entries 1024 name cgroup_ips
+load-map-cgroup_info_map:
+	[ -f $(PROG_MOUNT_PATH)/cgroup_info_map ] || sudo bpftool map create $(PROG_MOUNT_PATH)/cgroup_info_map type lru_hash key 8 value 24 entries 1024 name cgroup_info_map
 
 load-map-mark_pod_ips_map:
 	[ -f $(PROG_MOUNT_PATH)/mark_pod_ips_map ] || sudo bpftool map create $(PROG_MOUNT_PATH)/mark_pod_ips_map type hash key 4 value 16 entries 65535 name mark_pod_ips_map
@@ -114,7 +114,7 @@ clean-maps:
 		$(PROG_MOUNT_PATH)/sock_pair_map \
 		$(PROG_MOUNT_PATH)/pair_original_dst \
 		$(PROG_MOUNT_PATH)/process_ip \
-		$(PROG_MOUNT_PATH)/cgroup_ips \
+		$(PROG_MOUNT_PATH)/cgroup_info_map \
 		$(PROG_MOUNT_PATH)/local_pod_ips \
 		$(PROG_MOUNT_PATH)/cookie_original_dst \
 		$(PROG_MOUNT_PATH)/mark_pod_ips_map
@@ -141,13 +141,13 @@ clean-redir:
 	sudo bpftool prog detach pinned $(PROG_MOUNT_PATH)/redir msg_verdict pinned $(PROG_MOUNT_PATH)/sock_pair_map
 	sudo rm $(PROG_MOUNT_PATH)/redir
 
-load-connect: load-map-cookie_original_dst load-map-local_pod_ips load-map-process_ip load-map-cgroup_ips load-map-mark_pod_ips_map
+load-connect: load-map-cookie_original_dst load-map-local_pod_ips load-map-process_ip load-map-cgroup_info_map load-map-mark_pod_ips_map
 	sudo bpftool -m prog loadall mb_connect.o $(PROG_MOUNT_PATH)/connect \
 		map name cookie_original_dst pinned $(PROG_MOUNT_PATH)/cookie_original_dst \
 		map name local_pod_ips pinned $(PROG_MOUNT_PATH)/local_pod_ips \
 		map name mark_pod_ips_map pinned $(PROG_MOUNT_PATH)/mark_pod_ips_map \
 		map name process_ip pinned $(PROG_MOUNT_PATH)/process_ip \
-		map name cgroup_ips pinned $(PROG_MOUNT_PATH)/cgroup_ips
+		map name cgroup_info_map pinned $(PROG_MOUNT_PATH)/cgroup_info_map
 
 attach-connect:
 ifeq ($(ENABLE_IPV4),true)
@@ -196,9 +196,11 @@ ifeq ($(ENABLE_IPV4),true)
 	sudo rm -rf $(PROG_MOUNT_PATH)/bind
 endif
 
-load-sendmsg: load-map-cookie_original_dst
+load-sendmsg: load-map-cookie_original_dst load-map-cgroup_info_map load-map-mark_pod_ips_map
 	sudo bpftool -m prog loadall mb_sendmsg.o $(PROG_MOUNT_PATH)/sendmsg \
-		map name cookie_original_dst pinned $(PROG_MOUNT_PATH)/cookie_original_dst
+		map name cookie_original_dst pinned $(PROG_MOUNT_PATH)/cookie_original_dst \
+		map name mark_pod_ips_map pinned $(PROG_MOUNT_PATH)/mark_pod_ips_map \
+		map name cgroup_info_map pinned $(PROG_MOUNT_PATH)/cgroup_info_map
 
 attach-sendmsg:
 ifeq ($(ENABLE_IPV4),true)
@@ -217,9 +219,11 @@ ifeq ($(ENABLE_IPV6),true)
 endif
 	sudo rm -rf $(PROG_MOUNT_PATH)/sendmsg
 
-load-recvmsg: load-map-cookie_original_dst
+load-recvmsg: load-map-cookie_original_dst load-map-cgroup_info_map load-map-mark_pod_ips_map
 	sudo bpftool -m prog loadall mb_recvmsg.o $(PROG_MOUNT_PATH)/recvmsg \
-		map name cookie_original_dst pinned $(PROG_MOUNT_PATH)/cookie_original_dst
+		map name cookie_original_dst pinned $(PROG_MOUNT_PATH)/cookie_original_dst \
+		map name mark_pod_ips_map pinned $(PROG_MOUNT_PATH)/mark_pod_ips_map \
+		map name cgroup_info_map pinned $(PROG_MOUNT_PATH)/cgroup_info_map
 
 attach-recvmsg:
 ifeq ($(ENABLE_IPV4),true)