--- a/bpf/headers/mesh.h
+++ b/bpf/headers/mesh.h
@@ -13,7 +13,7 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
 */
-
+#pragma once
 #define SOCK_IP_MARK_PORT 39807
 
 #define ISTIO 1