bpf_sk_lookup_tcp may cost 50-150ns per ops, It is more expensive.

Without cache:
```
 kebe@Kebe-DaoCloud-PC  ~/workspace/mebpf   main ±  sudo go test -bench .
goos: linux
goarch: amd64
pkg: github.com/merbridge/merbridge
cpu: AMD Ryzen 9 3900 12-Core Processor             
BenchmarkConnect/eBPF-24                1000000000               0.4532 ns/op
BenchmarkConnect/eBPF/mb_sock_connect        108              3203 ns/op
BenchmarkSockops/eBPF-24                1000000000               0.4327 ns/op
BenchmarkSockops/eBPF/mb_sockops             954               689.2 ns/op
PASS
ok      github.com/merbridge/merbridge  20.530s
```

After cached:
```
 kebe@Kebe-DaoCloud-PC  ~/workspace/mebpf   main ±  sudo go test -bench .
goos: linux
goarch: amd64
pkg: github.com/merbridge/merbridge
cpu: AMD Ryzen 9 3900 12-Core Processor             
BenchmarkConnect/eBPF-24                1000000000               0.4539 ns/op
BenchmarkConnect/eBPF/mb_sock_connect        107              2475 ns/op
BenchmarkSockops/eBPF-24                1000000000               0.4545 ns/op
BenchmarkSockops/eBPF/mb_sockops             963               672.1 ns/op
PASS
ok      github.com/merbridge/merbridge  21.205s
```

The optimization results in a 20% performance improvement for connect programs.