@kebe7jun 2022-10-14
> Is it possible to optimize `is_port_listen_current_ns` as well?

Great, I am trying to optimize these calls in a unified way.